chore: move itinerary page mqtt management functions to utils, refactor
